20230212 임시커밋-사용

+++ PickaBoo/jam/src/main/java/com/funny/jam/dto/ViewCrawlingDataDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 public class ViewCrawlingDataDto { // 관리자 페이지에서 사용해야 할 부분
-    private Double image_index; // 인덱스(seq)
+    private Long image_index; // 인덱스(seq)
     private String image_Remote_path; // 이미지 경로
     private String image_kyeword; // 키워드
     private String image_category; // 카테고리 -enum
